[home] 전체 공지 화면 코드 merge
